services/qa_pipeline/requirements.txt
numpy>=1.22.0
scipy>=1.8.0
pydantic
fastapi
uvicorn[standard]
nats-py>=2.2.0
tomli>=2.0.1
